pyproject.toml
[tool.poetry]
name = "course7-drf"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "course7_drf"}]

[tool.poetry.dependencies]
python = "^3.11"
generics = "^6.0.0"
djangorestframework = "^3.14.0"
requests = "^2.31.0"
telebot = "^0.0.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
